+++ examples/hello-world/src/app/app.component.ts
@@ -11,13 +11,30 @@ import 'rxjs/add/operator/debounceTime';
 import 'rxjs/add/operator/map';
 import 'rxjs/add/operator/toPromise';
 
+interface User {
+  firstName: string;
+  lastName: string;
+  emails: {
+    address: string,
+    verified: boolean
+  }[];
+}
+
+interface GetUsersQueryResult {
+  users: User[];
+}
+
+interface AddUserMutationResult {
+  addUser: User;
+}
+
 @Component({
   selector: 'app-root',
   templateUrl: 'app.component.html',
 })
 export class AppComponent implements OnInit, AfterViewInit {
   // Observable with GraphQL result
-  public users: ApolloQueryObservable<any>;
+  public users: ApolloQueryObservable<GetUsersQueryResult>;
   public firstName: string;
   public lastName: string;
   public nameControl = new FormControl();
@@ -32,7 +49,7 @@ export class AppComponent implements OnInit, AfterViewInit {
 
   public ngOnInit() {
     // Query users data with observable variables
-    this.users = this.apollo.watchQuery({
+    this.users = this.apollo.watchQuery<GetUsersQueryResult>({
       query: gql`
         query getUsers($name: String) {
           users(name: $name) {
@@ -50,7 +67,7 @@ export class AppComponent implements OnInit, AfterViewInit {
       },
     })
       // Return only users, not the whole ApolloQueryResult
-      .map(result => result.data.users) as ApolloQueryObservable<any>;
+      .map(result => result.data.users) as any;
 
     // Add debounce time to wait 300 ms for a new change instead of keep hitting the server
     this.nameControl.valueChanges.debounceTime(300).subscribe(name => {
@@ -65,7 +82,7 @@ export class AppComponent implements OnInit, AfterViewInit {
 
   public newUser(firstName: string) {
     // Call the mutation called addUser
-    this.apollo.mutate({
+    this.apollo.mutate<AddUserMutationResult>({
       mutation: gql`
         mutation addUser(
           $firstName: String!
@@ -90,8 +107,8 @@ export class AppComponent implements OnInit, AfterViewInit {
       },
     })
       .toPromise()
-      .then(({ data }: ApolloQueryResult<any>) => {
-        console.log('got data', data);
+      .then(({ data }) => {
+        console.log('got a new user', data.addUser);
 
         // get new data
         this.users.refetch();