+++ src/index.ts
@@ -1,4 +1,5 @@
 export { SelectPipe } from './SelectPipe';
-export { Apollo } from './Apollo';
+export { Apollo, ApolloBase, provideClientMap } from './Apollo';
 export { ApolloQueryObservable } from './ApolloQueryObservable';
+export { CLIENT_MAP, CLIENT_MAP_WRAPPER } from './tokens';
 export { ApolloModule, APOLLO_PROVIDERS, APOLLO_DIRECTIVES, defaultApolloClient } from './ApolloModule';