+++ tests/ApolloModule.spec.ts
@@ -3,7 +3,7 @@ import { ApolloClient } from 'apollo-client';
 import './_common';
 
 import { ApolloModule, SelectPipe, Apollo } from '../src';
-import { APOLLO_CLIENT_WRAPPER, APOLLO_CLIENT_INSTANCE } from '../src/tokens';
+import { CLIENT_MAP, CLIENT_MAP_WRAPPER } from '../src/tokens';
 
 describe('ApolloModule', () => {
   const metadata: any = ApolloModule['decorators'][0]['args'][0];
@@ -35,17 +35,49 @@ describe('ApolloModule', () => {
     });
 
     it('should provide a wrapper directly', () => {
-      expect(providers[0]['provide']).toBe(APOLLO_CLIENT_WRAPPER);
+      expect(providers[0]['provide']).toBe(CLIENT_MAP_WRAPPER);
       expect(providers[0]['useValue']).toBe(getClient);
     });
 
     it('should provide a value using factory', () => {
       const factoryResult = providers[1]['useFactory'](getClient);
 
-      expect(providers[1]['provide']).toBe(APOLLO_CLIENT_INSTANCE);
+      expect(providers[1]['provide']).toBe(CLIENT_MAP);
       expect(providers[1]['useFactory']).toBeDefined();
-      expect(providers[1]['deps'][0]).toBe(APOLLO_CLIENT_WRAPPER);
+      expect(providers[1]['deps'][0]).toBe(CLIENT_MAP_WRAPPER);
       expect(factoryResult).toBe(client);
+    });
+  });
+
+  describe('forRoot', () => {
+    const defaultClient = {} as ApolloClient;
+    const extraClient = {} as ApolloClient;
+    const getClients = () => ({
+      default: defaultClient,
+      extra: extraClient,
+    });
+    const result = ApolloModule.forRoot(getClients);
+    const providers = result.providers[1]; // skips APOLLO_PROVIDERS
+
+    it('should contain ApolloModule as ngModule', () => {
+      expect(result.ngModule === ApolloModule).toBe(true);
+    });
+
+    it('should provide a wrapper directly', () => {
+      expect(providers[0]['provide']).toBe(CLIENT_MAP_WRAPPER);
+      expect(providers[0]['useValue']).toBe(getClients);
+    });
+
+    it('should provide a value using factory', () => {
+      const factoryResult = providers[1]['useFactory'](getClients);
+
+      expect(providers[1]['provide']).toBe(CLIENT_MAP);
+      expect(providers[1]['useFactory']).toBeDefined();
+      expect(providers[1]['deps'][0]).toBe(CLIENT_MAP_WRAPPER);
+      expect(factoryResult).toEqual({
+        default: defaultClient,
+        extra: extraClient,
+      });
     });
   });
 });